Reworked on animals, moved code from EntityAnimalTFC to IAnimalTFC for better addon compatibility. Fixed familiarity decaying every day instead of every second day.

+++ src/main/java/net/dries007/tfc/client/model/animal/ModelChickenTFC.java
@@ -83,7 +83,7 @@ public class ModelChickenTFC extends ModelBase
 
         EntityChickenTFC chicken = ((EntityChickenTFC) entity);
 
-        float percent = chicken.getPercentToAdulthood();
+        float percent = (float) chicken.getPercentToAdulthood();
         float ageScale = 2.0F - percent;
 
         GlStateManager.pushMatrix();
